+++ CHANGES.md
@@ -1,5 +1,7 @@
 * @dev
   * Added widget tooltip with current date
+  * Added option to override used locale
+  * Added option to set custom week start day
 
 * v1.1.0 (2020-12-16)
   * General code cleamup